+++ src/Registry/Source/SourceInterface.php
@@ -2,7 +2,7 @@
 /**
  *	...
  *
- *	Copyright (c) 2016-2020 Christian Würker (ceusmedia.de)
+ *	Copyright (c) 2016-2023 Christian Würker (ceusmedia.de)
  *
  *	This program is free software: you can redistribute it and/or modify
  *	it under the terms of the GNU General Public License as published by
@@ -20,8 +20,8 @@
  *	@category		Library
  *	@author			Christian Würker <christian.wuerker@ceusmedia.de>
  *	@package		CeusMedia_Router_Registry_Source
- *	@copyright		2016-2020 Christian Würker
- *	@license		http://www.gnu.org/licenses/gpl-3.0.txt GPL 3
+ *	@copyright		2016-2023 Christian Würker
+ *	@license		https://www.gnu.org/licenses/gpl-3.0.txt GPL 3
  *	@link			https://github.com/CeusMedia/Router
  */
 
@@ -35,8 +35,8 @@ use CeusMedia\Router\Registry;
  *	@category		Library
  *	@package		CeusMedia_Router_Registry_Source
  *	@author			Christian Würker <christian.wuerker@ceusmedia.de>
- *	@copyright		2016-2020 Christian Würker
- *	@license		http://www.gnu.org/licenses/gpl-3.0.txt GPL 3
+ *	@copyright		2016-2023 Christian Würker
+ *	@license		https://www.gnu.org/licenses/gpl-3.0.txt GPL 3
  *	@link			https://github.com/CeusMedia/Router
  */
 interface SourceInterface
@@ -46,9 +46,9 @@ interface SourceInterface
 
 	/**
 	 *	@param		string|NULL		$resource
-	 *	@return		AbstractSource
+	 *	@return		SourceInterface
 	 */
-	public static function create( string $resource = NULL ): AbstractSource;
+	public static function create( string $resource = NULL ): SourceInterface;
 
 	/**
 	 *	@param		string|NULL		$resource
@@ -59,7 +59,7 @@ interface SourceInterface
 	 *	@param		int			$key
 	 *	@return		mixed|NULL
 	 */
-	public function getOption( int $key );
+	public function getOption( int $key ): mixed;
 
 	/**
 	 *	@return		string|NULL
@@ -77,7 +77,7 @@ interface SourceInterface
 	 *	@param		mixed|NULL		$value
 	 *	@return		mixed
 	 */
-	public function setOption( int $key, $value = NULL );
+	public function setOption( int $key, mixed $value = NULL ): mixed;
 
 	/**
 	 *	@param		string			$resource