+++ test/RouterTest.php
@@ -1,10 +1,10 @@
 <?php
+namespace CeusMedia\RouterTest;
+
 use PHPUnit\Framework\TestCase;
 use CeusMedia\Router\Log;
 use CeusMedia\Router\Router;
-use CeusMedia\Router\Route;
 use CeusMedia\Router\ResolverException;
-use CeusMedia\Router\Route\Factory as RouteFactory;
 use CeusMedia\Router\Registry\Source\JsonFile as JsonFileRegistry;
 
 /**
@@ -12,8 +12,6 @@ use CeusMedia\Router\Registry\Source\JsonFile as JsonFileRegistry;
  */
 class RouterTest extends TestCase
 {
-	protected $factory;
-
 	protected function setUp(): void
 	{
 	}
@@ -23,10 +21,11 @@ class RouterTest extends TestCase
 	}
 
 	/**
-	 *	@covers	::resolve
-	 *	@covers	\CeusMedia\Router\Resolver::resolve
+	 * @covers    ::resolve
+	 * @covers    \CeusMedia\Router\Resolver::resolve
+	 * @throws ResolverException
 	 */
-	public function testResolve()
+	public function testResolve(): void
 	{
 		$jsonFile	= __DIR__.'/JsonFileTest.routes.json';
 		$router	= new Router();
@@ -39,33 +38,33 @@ class RouterTest extends TestCase
 		//  this CLI test fails -> argument matching for CLI is not separately handled
 //		$route	= $routes[array_keys( $routes )[0]];
 //		$actual	= $router->resolve( 'test a1 b2' );
-//		$this->assertEquals( $route, $actual );
+//		self::assertEquals( $route, $actual );
 
 		$router->setMethod( 'GET' );
 
 		$route	= $routes[array_keys( $routes )[1]];
 		$actual	= $router->resolve( '/test' );
-		$this->assertEquals( $route, $actual );
+		self::assertEquals( $route, $actual );
 
 		$route	= $routes[array_keys( $routes )[2]];
 		$route->setArguments( ['a' => 'a1'] );
 		$actual	= $router->resolve( '/test/a1' );
-		$this->assertEquals( $route, $actual );
+		self::assertEquals( $route, $actual );
 
 		$route	= $routes[array_keys( $routes )[3]];
 		$route->setArguments( ['a' => 'a1', 'b' => 'b2'] );
 		$actual	= $router->resolve( '/test/a1/b2' );
-		$this->assertEquals( $route, $actual );
+		self::assertEquals( $route, $actual );
 
 		$actual	= $router->resolve( '/test/a1/b2/c3', FALSE );
-		$this->assertEquals( NULL, $actual );
+		self::assertEquals( NULL, $actual );
 	}
 
 	/**
 	 *	@covers	::resolve
 	 *	@covers	\CeusMedia\Router\Resolver::resolve
 	 */
-	public function testResolveException()
+	public function testResolveException(): void
 	{
 		$this->expectException( ResolverException::class );
 		$jsonFile	= __DIR__.'/JsonFileTest.routes.json';
@@ -78,7 +77,7 @@ class RouterTest extends TestCase
 		//  this CLI test fails -> argument matching for CLI is not separately handled
 //		$route	= $routes[array_keys( $routes )[0]];
 //		$actual	= $router->resolve( 'test a1 b2 c3' );
-//		$this->assertEquals( $route, $actual );
+//		self::assertEquals( $route, $actual );
 
 		$router->setMethod( 'GET' );
 		$router->resolve( '/test/a1/b2/c3' );