+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog for CeusMedia/Router
 
+## Version 0.5
+- PHP 8.1 support and migration.
+- Migrate to support CeusMedia/Common 1.0.x.
+- Use PHPUnit 10.1.
+- Reach maximum PHPStan level.
+- Move tool config files to util.
+
 ## Version 0.4
 - Extend route model by roles.
 - Migrate to support CeusMedia/Common 0.9.x